fix: use global reservation interal

+++ core/mock/payment_state.go
@@ -62,9 +62,9 @@ func (m *MockOnchainPaymentState) GetGlobalSymbolsPerSecond() uint64 {
 	return args.Get(0).(uint64)
 }
 
-func (m *MockOnchainPaymentState) GetGlobalRateBinInterval() uint64 {
+func (m *MockOnchainPaymentState) GetGlobalRateBinInterval() uint32 {
 	args := m.Called()
-	return args.Get(0).(uint64)
+	return args.Get(0).(uint32)
 }
 
 func (m *MockOnchainPaymentState) GetMinNumSymbols() uint32 {
